Revamp JPG data model +bug fixes
- Revamp JPG DM to make use of ModelHelper().
- Fix INT.absorb_auto_helper() that was creating an int_list with one value when
  the INT did not have an int_list but only min and max. The side effect is that
  when INT.do_absorb() is called, it will make the INT having an int_list, even if it first didn't
  have one. Thus, cloning such INT even by ignoring its frozen state will never provide
  the original INT. (which may stop mistakenly the absorption of components with a multiplicity>1).
- Fix LocalTarget.cleanup().

+++ framework/target.py
@@ -1516,6 +1516,9 @@ class LocalTarget(Target):
         self._data_sent = True
         
     def cleanup(self):
+        if self.__app is None:
+            return
+
         try:
             os.kill(self.__app.pid, signal.SIGTERM)
         except:
